fix: fill one node from several
sometimes when fixing you need to fill one node for several
of its right siblings.
(i assumed filling from the next sibling was always enough but it's
not).
so we cannot loop as easily as before on all siblings because we now
have a more complex state with the node to fill and the next node to
fill from